GRA-30// Rename register.fxml -> MainScene.fxml .
