+++ src/main/resources/MainScene.fxml
@@ -9,13 +9,18 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
 
-<BorderPane fx:id="borderPane" prefHeight="1000.0" prefWidth="1400.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="rep.PP.Controller">
+<BorderPane fx:id="borderPane" prefHeight="1000.0" prefWidth="1400.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="rep.PP.controllers.Controller">
    <center>
       <VBox alignment="CENTER" style="-fx-background-color: #202020;" styleClass="root" stylesheets="@style.css" BorderPane.alignment="CENTER">
          <children>
-            <AnchorPane cache="true" cacheHint="QUALITY" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="980.0" prefWidth="1400.0" styleClass="back" VBox.vgrow="ALWAYS" />
+            <StackPane VBox.vgrow="ALWAYS">
+               <children>
+                  <AnchorPane cache="true" cacheHint="QUALITY" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="980.0" prefWidth="1400.0" styleClass="back" />
+               </children>
+            </StackPane>
          </children>
       </VBox>
    </center>
