--- src/main/resources/register.fxml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.effect.Light.Distant?>
-<?import javafx.scene.effect.Lighting?>
-<?import javafx.scene.effect.Shadow?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.VBox?>
-
-<BorderPane fx:id="borderPane" prefHeight="1000.0" prefWidth="1400.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="rep.PP.Controller">
-   <center>
-      <VBox alignment="CENTER" style="-fx-background-color: #202020;" styleClass="root" stylesheets="@style.css" BorderPane.alignment="CENTER">
-         <children>
-            <AnchorPane cache="true" cacheHint="QUALITY" maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="980.0" prefWidth="1400.0" styleClass="back" VBox.vgrow="ALWAYS" />
-         </children>
-      </VBox>
-   </center>
-   <left>
-      <VBox prefHeight="1029.0" prefWidth="338.0" style="-fx-background-color: #202020;" BorderPane.alignment="CENTER" />
-   </left>
-   <effect>
-      <Lighting diffuseConstant="1.51">
-         <bumpInput>
-            <Shadow height="138.0" radius="39.19" width="20.76" />
-         </bumpInput>
-         <light>
-            <Light.Distant />
-         </light>
-      </Lighting>
-   </effect>
-   <top>
-      <HBox alignment="CENTER_RIGHT" onMouseDragged="#dragged" onMousePressed="#pressed" style="-fx-background-color: #a16209;">
-         <children>
-            <ImageView fitHeight="24.0" fitWidth="24.0" onMouseClicked="#MIN" pickOnBounds="true" preserveRatio="true">
-               <HBox.margin>
-                  <Insets left="10.0" right="10.0" />
-               </HBox.margin>
-               <image>
-                  <Image url="@MinimizeIcon.png" />
-               </image>
-            </ImageView>
-            <ImageView fitHeight="24.0" fitWidth="24.0" onMouseClicked="#MAX" pickOnBounds="true" preserveRatio="true">
-               <HBox.margin>
-                  <Insets left="10.0" right="10.0" />
-               </HBox.margin>
-               <image>
-                  <Image url="@FullIcon.png" />
-               </image>
-            </ImageView>
-            <ImageView fx:id="closebutton" accessibleRole="BUTTON" fitHeight="24.0" fitWidth="24.0" onMouseClicked="#close" pickOnBounds="true" preserveRatio="true">
-               <image>
-                  <Image url="@CloseIcon.png" />
-               </image>
-               <HBox.margin>
-                  <Insets left="10.0" right="10.0" />
-               </HBox.margin>
-            </ImageView>
-         </children>
-         <effect>
-            <Lighting diffuseConstant="1.51" specularConstant="0.32">
-               <bumpInput>
-                  <Shadow />
-               </bumpInput>
-               <light>
-                  <Light.Distant />
-               </light>
-            </Lighting>
-         </effect>
-      </HBox>
-   </top>
-</BorderPane>
